refactor: 임베딩 관련 모듈 리팩토링 및 구조 개선 (#61)
- `embedder.py` 제거 및 임베딩 기능을 `jina_embeddings_v3.py`로 통합
- 모든 참조를 `src.vector_db.embedder`에서 `src.models.jina_embeddings_v3.embed`로 교체
- 임베딩 로직을 클래스(`JinaEmbeddingFunction`) 기반으로 구현 및 재사용성 개선 (#61)
