+++ src/models/jina_embeddings_v3.py
@@ -1,35 +1,47 @@
 import torch
-from src.core.ai_logger import get_ai_logger
+from typing import List, Union
 from sentence_transformers import SentenceTransformer
-
-# 초기화 시 한번만 호출
-model = model = SentenceTransformer("jinaai/jina-embeddings-v3", trust_remote_code=True)
+from chromadb.api.types import EmbeddingFunction
+from src.config import EMBED_MODEL
+from src.core.ai_logger import get_ai_logger
 
 ai_logger = get_ai_logger()
+_MODEL = SentenceTransformer(EMBED_MODEL, trust_remote_code=True).to("cpu")
 
-device = "cuda" if torch.cuda.is_available() else "cpu"
+class JinaEmbeddingFunction(EmbeddingFunction):
+    def __init__(self, model_name: str = EMBED_MODEL):
+        self._model_name = model_name
+        self.device = "cpu"
+        self.model = _MODEL
+        ai_logger.info(f"[AI] [임베딩 모델 로드 완료]: {model_name} ({self.device})")
 
-def embed(texts: list[str] | str) -> list[list[float]]:
-    if isinstance(texts, str):
-        texts = [texts]
-    elif not isinstance(texts, list):
-        raise ValueError("embed 함수는 문자열 또는 문자열 리스트만 지원합니다.")
+    def __call__(self, input: Union[List[str], str]) -> List[List[float]]:
+        if isinstance(input, str):
+            input = [input]
+        elif not isinstance(input, list):
+            raise ValueError("입력은 문자열 또는 문자열 리스트여야 합니다.")
 
-    try:
-        ai_logger.info("[AI] [임베딩 요청]", extra={"input_count": len(texts)})
+        try:
+            ai_logger.info("[AI] [임베딩 요청]", extra={"input_count": len(input)})
 
-        # 임베딩 수행
-        vectors = model.encode(texts, convert_to_numpy=True).tolist()
+            vectors = self.model.encode(input, convert_to_numpy=True).tolist()
 
-        if len(vectors) != len(texts):
-            ai_logger.warning("[AI] [임베딩 수 불일치]", extra={"input_count": len(texts), "output_count": len(vectors)})
+            if len(vectors) != len(input):
+                ai_logger.warning("[AI] [임베딩 수 불일치]", extra={
+                    "input_count": len(input), "output_count": len(vectors)
+                })
 
-        ai_logger.info("[AI] [임베딩 완료]", extra={"vector_dim": len(vectors[0]) if vectors else 0})
-        return vectors
+            ai_logger.info("[AI] [임베딩 완료]", extra={"vector_dim": len(vectors[0]) if vectors else 0})
+            return vectors
 
-    except Exception:
-        ai_logger.exception("[AI] [임베딩 실패]")
-        return []
+        except Exception:
+            ai_logger.exception("[AI] [임베딩 실패]")
+            return []
+
+    def name(self) -> str:
+        return self._model_name
+
+embed = JinaEmbeddingFunction()
 
 if __name__ == "__main__":
-    print("[임베딩 차원 확인]", len(model.encode(["test"], convert_to_numpy=True)[0]))
+    print("[임베딩 차원 확인]", len(embed(["test"])[0]))
