+++ src/vector_db/user_document_builder.py
@@ -1,5 +1,5 @@
 from src.vector_db.chroma_client import get_chroma_client
-from src.vector_db.embedder import JinaEmbeddingFunction
+from src.models.jina_embeddings_v3 import embed
 
 def build_user_document(user_id: str, group_id: list[str]) -> list[dict]:
     if not user_id or not group_id:
@@ -19,7 +19,6 @@ def build_user_document(user_id: str, group_id: list[str]) -> list[dict]:
 
 def remove_user_group_activity(user_id: str, group_id: str) -> None:
     client = get_chroma_client()
-    embed = JinaEmbeddingFunction()
     col = client.get_or_create_collection(name="user-activity", embedding_function=embed)
     doc_id = f"user-{user_id}-{group_id}"
     col.delete(ids=[doc_id])
